refactor: remove legacy bridge container
